Rename folder nsis-streambox2 to nsisscripts
This folder is very much a stripped down copy of nsis-streambox2.  Its
purpose is not the same as nsis-streambox2.  nsis-streambox2 going
forward is meant to be deployed once on development machine and then nsi
installers will pull files it need from ${NSISDIR}\Stremabox folder.

nsisscripts is meant for versioned copies of .nsh and .nsi libraries.
Its not meant to be put in ${NSISDIR}\Streambox folder.